mcom: Rename variables $draft and $draftmime
$draft is renamed to $msg_sketch
$draftmime is renamed to $draft

I'm doing this as a preparation for a rework on the gpg handling.
For safe gpg handling, it is necessary to strictly separate the
local 'sketch' from the 'draft'. The 'draft' should be a valid mail,
might be synchronized with a remote IMAP server and should therefore
already be encrypted, when gpg is used.

This commit doesn't touch any logic. Further commits will follow,
which ensure that the 'sketch' is always transformed into a 'draft'
at some point.