mcom: Refactor to prepare for safer gpg handling
With this commit, two new functions are introduced:
- sketch_to_draft() and
- draft_to_sketch()

These two functions will - in later commits - be extended to handle
the encrption and signing, so that no unencrypted draft will ever
lie in the outbox, when gpg encryption is used. $msg_sketch is now
a file in /tmp and will always be removed when mcom exits.

The introduction of the two new functions required quite some
refactoring and even slight changes in the logic. $msg_sketch is
not passed to $sendmail anymore. Instead $msg_sketch will *always*
be transformed into $draft, before sending or aborting and only
$draft will ever be passed to $sendmail.

Consequently the $automime variable, as well as the whole 'mime'
option could be removed. The do_mime() function will always be called,
when $msg_sketch is transformed to $draft via sketch_to_draft().