mrep: fix subject not populating when pager stdout is not tty
If the pager is set to something where stdout is not a TTY, such as a
text editor, it causes non-interactive calls to mscan to fail with the
error: `Fatal error: stdout is not a tty`. When building the subject,
`mrep` fetches the subject from the email being replied to using
`mscan`, causing this error to manifest and the subject line to end up
as: "Re: ". Explicitly setting the pager to `cat` when fetching the
email subject in `mrep` prevents this error.